WIP: Add testing for Fable.Import.Node
Currently, the node bindings do not have any tests to ensure correctness.

This makes it cumbersome to work with changes, as you need to manually test them.

Instead we can add tests, that ensure the bindings are working as intended, and catch any regressions going forward.

This patch sets up a test directory (unsure if that's the best place to start something like this, or they should be nested in each bindings dir), and adds some tests for Node.Streams.

I have not added a travis / appveyor file, but we should be able to use the build pipeline feature of travis to have multiple parallel jobs  .

I am also using Fable.Import.Jest for testing, but would be happy to switch to another runner if preferred.