Update testing to reflect multiple projects
The latest version of fable-jest supports
running with multiple subprojects.

Add a new fsproj for the node tests specifically,
and update the jest config / package.json to
reflect newer changes.

We still need a fsproj at the top test level,
but that's only because it seems to be necessary
to run the fable clitool there.